Agora-ElevenLabs-ConvoAI-Agent/requirements.txt
agora-realtime-ai-api==1.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiohttp[speedups]
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
black==24.4.2
certifi==2024.7.4
cffi==1.17.1
click==8.1.7
colorlog>=6.0.0
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
mypy==1.10.1
mypy-extensions==1.0.0
numpy==1.26.4
numpy>=1.21.0
# openai==1.37.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psutil==5.9.8
protobuf==5.27.2
PyAudio==0.2.14
pyaudio>=0.2.11
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyee==12.0.0
PyJWT==2.8.0
pytest==8.2.2
python-dotenv==1.0.1
ruff==0.5.2
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
sounddevice>=0.4.6
tqdm==4.66.4
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
watchfiles==0.22.0
yarl==1.12.1
elevenlabs>=0.3.0
scipy==1.11.3

